requirements.txt
absl-py==1.0.0
altair==4.2.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
asttokens==2.0.5
attrs==21.4.0
autolab-core==1.1.0
av==8.1.0
backcall==0.2.0
base58==2.1.1
black==21.4b2
bleach==4.1.0
blinker==1.4
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
cloudpickle==2.0.0
colorlog==6.6.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
detectron2==0.6+cu102
dill==0.3.4
entrypoints==0.3
executing==0.8.2
fonttools==4.28.5
freetype-py==2.2.0
future==0.18.2
fvcore==0.1.5.post20211023
gitdb==4.0.9
GitPython==3.1.26
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.43.0
hydra-core==1.1.1
idna==3.3
imageio==2.13.5
importlib-metadata==4.10.1
iopath==0.1.9
ipykernel==6.7.0
ipython==8.0.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.1.1
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lxml==4.7.1
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mpmath==1.2.1
msgpack==1.0.3
multiprocess==0.70.12.2
mypy-extensions==0.4.3
nbclient==0.5.10
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
notebook==6.4.7
numpy==1.22.1
oauthlib==3.1.1
omegaconf==2.1.1
opencv-python==4.5.5.62
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parameterized==0.8.1
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
portalocker==2.3.2
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.3
ptyprocess==0.7.0
pure-eval==0.2.1
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycollada==0.7.2
pycparser==2.21
pydeck==0.7.1
pydot==1.4.2
pyglet==1.5.21
Pygments==2.11.2
Pympler==1.0.1
PyOpenGL==3.1.0
pyparsing==3.0.6
pyrender==0.1.45
pyrsistent==0.18.1
python-dateutil==2.8.2
pytorchvideo==0.1.3
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
regex==2021.11.10
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
Rtree==0.9.7
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.0
setproctitle==1.2.2
Shapely==1.8.0
six==1.16.0
smmap==5.0.0
stack-data==0.1.4
streamlit==1.4.0
svg.path==4.1
sympy==1.9
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tifffile==2021.11.2
toml==0.10.2
toolz==0.11.2
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
trimesh==3.9.42
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
validators==0.18.2
visualization==1.0.0
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
xxhash==2.0.2
yacs==0.1.8
zipp==3.7.0
